UPSTREAM: drm/docs: Explicitly document default CRTC background behavior
Add a paragraph explaining that the default behavior for areas which
are not covered by planes or where planes are blending with the CRTC
background, is black.

This is alluded to in the "pixel blend mode" property docs, but not
called out explicitly.

Reviewed-by: Simon Ser <contact@emersion.fr>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20230103152738.1213785-1-sean@poorly.run
(cherry picked from commit 51342cc021400841b461cc579f76db24cdb482fc)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:280653303
TEST=build test

Backported using forklift.py

Change-Id: I026094337284b1732c6467ea9e97b8f1550cec69