CHROMIUM: iwl7000: mvm: set the firmware into TWT testmode if needed
We need to tell the firmware to be in TWT testmode on the new API (mld)
as well.
Not doing that allowed the firmware to trigger TWT handshake based on
its own decision, failing the WFA tests.

Since this patch is NOUPSTREAM, I prefer not to change the copyright in
this patch so that the next patch that will go upstream will change it
and then, the copyright will be updated upstream as well.

Change-Id: I896ba1884f37776b4c37825e85c7c2cb2041a6b3
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 1c7da892eca75d1f17dfa4aa47e6380242844aab