UPSTREAM: habanalabs/gaudi2: find decode error root cause
When a decode error happens, we often don't know the exact root
cause (the erroneous address that was accessed) and the exact engine
that created the erroneous transaction.

To find out, we need to go over all the relevant register blocks
in the ASIC. Once we find the relevant engine, we print its details
and the offending address.

This helps tremendously when debugging an error that was created
by running a user workload.

Signed-off-by: Koby Elbaz <kelbaz@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit f7d67c1cfdccfe7168d28c26b935c9da18bfdb8c)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:300094055
TEST=build test

Backported using forklift.py

Change-Id: Ibf2ccd83a14769439fb1ead69d5aab6e0262c9e4