UPSTREAM: accel/habanalabs: add info when FD released while device still in use
When user closes the device file descriptor, it is checked whether the
device is still in use, and a message is printed if it is.
To make this message more informative, add to this print also the reason
due to which the device is considered as in use.
The possible reasons which are checked for now are active CS and
exported dma-buf.

Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
Reviewed-by: Stanislaw Gruszka <stanislaw.gruszka@linux.intel.com>
(cherry picked from commit d43bce6e762f25b25685487630510452feaf7362)

BUG=b:300094055
TEST=build test

Change-Id: Ia099c010e9ff7c30731fb43ded4fcfb639c269b0
Signed-off-by: Drew Davenport <ddavenport@chromium.org>