UPSTREAM: gpu: host1x: Implement syncpoint wait using DMA fences
In anticipation of removal of the intr API, move host1x_syncpt_wait
to use DMA fences instead. As of this patch, this means that waits
have a 30 second maximum timeout because of the implicit timeout
we have with fences, but that will be lifted in a follow-up patch.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit f0fb260a0cdb014b22a5f7733279c205f2cba62a)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:280653303
TEST=build test

Backported using forklift.py

Change-Id: Iae030c03c42da640e406fc5e6f46467bfd52d3f8