FROMGIT: drm/i915/pxp: Invalidate all PXP fw sessions during teardown
A gap was recently discovered where if an application did not
invalidate all of the stream keys (intentionally or not), and the
driver did a full PXP global teardown on the GT subsystem, we
find that future session creation would fail on the security
firmware's side of the equation. i915 is the entity that needs
ensure the sessions' state across both iGT and security firmware
are at a known clean point when performing a full global teardown.

Architecturally speaking, i915 should inspect all active sessions
and submit the invalidate-stream-key PXP command to the security
firmware for each of them. However, for the upstream i915 driver
we only support the arbitration session that can be created
so that will be the only session we will cleanup.

Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Reviewed-by: Juston Li <justonli@chromium.org>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230125082637.118970-5-alan.previn.teres.alexis@intel.com
(cherry picked from commit d374c047b38e9f1130308aae207dc44045cd5cac
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: Ib57cb6f47b1eada272706c294d43b551fe44f613
Signed-off-by: Rakshith M O <rakshith.m.o@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4280850
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Waheem Md <waheem.md@intel.com>
Commit-Queue: Paz Zcharya <pazz@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>