UPSTREAM: thermal/core: Add a thermal zone 'devdata' accessor
The thermal zone device structure is exposed to the different drivers
and obviously they access the internals while that should be
restricted to the core thermal code.

In order to self-encapsulate the thermal core code, we need to prevent
the drivers accessing directly the thermal zone structure and provide
accessor functions to deal with.

Provide an accessor to the 'devdata' structure and make use of it in
the different drivers.

No functional changes intended.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit a6ff3c0021468721b96e84892a8cae24bde8d65f)

BUG=b:236331727
BUG=b:326555352
TEST=emerge-geralt chromeos-kernel-6_1

Change-Id: Ia8312c2e08fbeb0ae7d598f7998dd12eff552b3b
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5321178
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>